refactor(gateway): migrate CapabilityRecord + slug helpers to gateway-core (#845)
Second migration in the #845 chain. Moves the gateway capability domain
type and its slug encoding contract into dcc-mcp-gateway-core so any
external Rust tooling that talks to the gateway over REST can pull just
the domain crate instead of the full gateway crate.

Migrated:
- CapabilityRecord (REST wire shape for POST /v1/search) + new(),
  from_skill_tool(), MAX_SUMMARY_LEN, normalise_summary helper.
- tool_slug() / parse_slug() / is_valid_dcc_bucket() — slug encoding
  and parsing helpers.
- SCHEMA_AVAILABLE annotation key.
- naming::instance_short() / naming::is_cursor_safe_alphabet() /
  naming::ID_PREFIX_LEN — primitives the slug helpers depend on.

dcc-mcp-gateway side:
- gateway/capability/record.rs is now a re-export shim so the
  historical paths (crate::gateway::capability::record::CapabilityRecord
  etc.) keep compiling unchanged across the 8-file capability subsystem,
  the 3 builder/refresh/search call sites, and the gateway tests.
- gateway/namespace/constants.rs forwards instance_short() and
  ID_PREFIX_LEN; gateway/namespace/encode.rs forwards
  is_cursor_safe_alphabet(). The full namespace module's bare-name
  parsing and resource-URI helpers stay in the gateway crate (they
  depend on dcc-mcp-naming and broader runtime state).

Cargo.toml changes for dcc-mcp-gateway-core:
- serde becomes a hard dependency (CapabilityRecord is wire-shaped).
- uuid added (the slug encoding rounds UUIDs to 8 hex chars).

Adds 2 new tests in dcc-mcp-gateway-core:
- from_skill_tool_uses_nil_uuid: documents the unloaded-skill sentinel.
- record_roundtrip_json_preserves_wire_shape: regression guard for the
  REST contract, including the empty-tags skip_serializing_if.

Verification:
- cargo check --workspace clean
- cargo test -p dcc-mcp-gateway-core: 16 passed
- cargo test -p dcc-mcp-gateway --lib capability::: 58 passed
- cargo test -p dcc-mcp-gateway --lib namespace::: 42 passed
- cargo clippy -p dcc-mcp-gateway --no-deps --tests --all-features
  -- -D warnings clean
- cargo fmt --all -- --check clean

Refs: #845, builds on #893/#894. Part of #848 epic.
