+++ crates/dcc-mcp-gateway-core/Cargo.toml
@@ -15,14 +15,10 @@ categories = ["api-bindings"]
 workspace = true
 
 [dependencies]
-serde = { workspace = true, optional = true }
-
-[features]
-default = []
-# Opt-in `serde::{Serialize, Deserialize}` derives for the domain types.
-# Kept off by default so downstream infra crates can pull the domain layer
-# without paying for serde codegen when they do not need it.
-serde = ["dep:serde"]
+# Domain types are part of the REST wire contract (`POST /v1/search`
+# returns CapabilityRecord arrays), so serde is a hard dependency.
+serde = { workspace = true }
+uuid = { workspace = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
